% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{data_setup}
\alias{data_setup}
\title{Load data from DB into Global Options}
\usage{
data_setup(pool)
}
\arguments{
\item{pool}{The database connection pool}
}
\description{
Load data from DB into Global Options
}
