% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_manage_users.R, R/fct_user_editor.R
\name{list_users}
\alias{list_users}
\title{List All Users With Roles}
\usage{
list_users(pool)

list_users(pool)
}
\arguments{
\item{pool}{A DBI connection pool.}
}
\value{
A data frame of all users and their roles.

A data frame of all users and their roles.
}
\description{
Returns a joined table of:
- usernames
- names
- email
- active flag
- expire_date
- created_at timestamp
- assigned PostgreSQL roles

Returns a joined table of:
- usernames
- names
- email
- active flag
- expire_date
- created_at timestamp
- assigned PostgreSQL roles
}
