% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_manage_users.R
\name{update_user}
\alias{update_user}
\title{Update an Existing User}
\usage{
update_user(
  pool,
  username,
  password = NULL,
  pg_role = NULL,
  active = NULL,
  first_name = NULL,
  last_name = NULL,
  email = NULL,
  expire_date = NULL,
  add_project_ids = NULL,
  remove_project_ids = NULL
)
}
\arguments{
\item{pool}{A DBI connection pool.}

\item{username}{Character. Username whose record is updated.}

\item{password}{Optional character. New password.}

\item{pg_role}{Optional character. New PostgreSQL role to add.}

\item{active}{Optional logical.}

\item{first_name}{Optional character.}

\item{last_name}{Optional character.}

\item{email}{Optional character.}

\item{expire_date}{Optional Date.}

\item{add_project_ids}{Optional integer vector. Project IDs to assign to the user.}

\item{remove_project_ids}{Optional integer vector. Project IDs to remove from the user.}
}
\value{
Invisible list with success status.
}
\description{
Allows updating of:
- password (rehashes automatically)
- active status
- first/last name
- email
- expire_date
- roles (adds new role; does not remove existing roles)
- projects (add and/or remove project assignments)
}
