% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_file_info_upsert.R
\name{upsert_audio_files_df}
\alias{upsert_audio_files_df}
\title{Upsert audio files into import.audio_files from a data.frame}
\usage{
upsert_audio_files_df(conn, df_audio, update_if_exists = TRUE)
}
\arguments{
\item{conn}{A valid DBI pool connection}

\item{df_audio}{data.frame with columns:
deployment_id, sample_rate, relative_path, file_deleted (logical),
timestamp_start (POSIXct or character), duration_s (integer),
audio_file_id (optional, for updates)}

\item{update_if_exists}{Logical. If TRUE, will update rows with audio_file_id.}
}
\value{
A vector of audio_file_id for inserted/updated rows
}
\description{
Upsert audio files into import.audio_files from a data.frame
}
