% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_file_info_upsert.R
\name{upsert_deployments_sf}
\alias{upsert_deployments_sf}
\title{Upsert deployments from an sf object}
\usage{
upsert_deployments_sf(conn, sf_deployments, update_if_exists = TRUE)
}
\arguments{
\item{conn}{A valid DBI pool connection}

\item{sf_deployments}{An sf object with columns matching import.deployments:
project_id, deployment_name, deployment_path, start_datetime, end_datetime,
device_manufacturer, device_modelname, valid, notes, deployment_id (optional)
and a geometry column.}

\item{update_if_exists}{Logical. If TRUE, will update rows where deployment_id exists.}
}
\value{
A vector of deployment_ids for inserted/updated rows
}
\description{
Upsert deployments from an sf object
}
