% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_file_info_upsert.R
\name{upsert_project}
\alias{upsert_project}
\title{Upsert a project into the import.projects table}
\usage{
upsert_project(
  conn,
  project_name_short,
  project_name_long,
  description = NULL,
  contact = NULL,
  organisation = NULL,
  project_id = NULL
)
}
\arguments{
\item{conn}{A valid DBI pool connection}

\item{project_name_short}{Short project name (max 30 chars)}

\item{project_name_long}{Long project name (max 150 chars)}

\item{description}{Text description}

\item{contact}{Contact info}

\item{organisation}{Organisation name}

\item{project_id}{Optional. If provided, will update that project; if NULL, will insert new}
}
\value{
The project_id of the inserted or updated project
}
\description{
Upsert a project into the import.projects table
}
