% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_file_info_upsert.R
\name{upsert_results_df}
\alias{upsert_results_df}
\title{Upsert results into import.results from a data.frame}
\usage{
upsert_results_df(conn, df_results, update_if_exists = TRUE)
}
\arguments{
\item{conn}{A valid DBI pool connection}

\item{df_results}{data.frame with columns:
audio_file_id, settings_id, begin_time_ms, end_time_ms,
confidence, species_id, behavior_id}

\item{update_if_exists}{Logical. If TRUE, will update rows matching unique constraint}
}
\value{
A vector of result_id for inserted/updated rows
}
\description{
Upsert results into import.results from a data.frame
}
