% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_file_info_upsert.R
\name{upsert_settings_df}
\alias{upsert_settings_df}
\title{Upsert settings into import.settings from a data.frame}
\usage{
upsert_settings_df(conn, df_settings, update_if_exists = TRUE)
}
\arguments{
\item{conn}{A valid DBI pool connection}

\item{df_settings}{data.frame with columns:
model_name, model_version (optional), min_conf, overlap,
locale (optional), model_params (list or character/JSON string)}

\item{update_if_exists}{Logical. If TRUE, will update existing rows on conflict.
Note: Since there's no UNIQUE constraint, this will always insert.}
}
\value{
A vector of settings_id for inserted rows
}
\description{
Upsert settings into import.settings from a data.frame
}
