% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{The application Server-Side Logic}
\usage{
app_server(input, output, session, pool)
}
\arguments{
\item{input}{Internal parameter for `{shiny}`.}

\item{output}{Internal parameter for `{shiny}`.}

\item{session}{Internal parameter for `{shiny}`.}

\item{pool}{The database connection pool.}
}
\description{
The application Server-Side Logic
}
