% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_birdnet_process_deployment.R
\name{process_deployment_birdnet}
\alias{process_deployment_birdnet}
\title{Process a single deployment with BirdNET inference}
\usage{
process_deployment_birdnet(
  deployment_id,
  deployments,
  audio_files,
  species,
  temporal_filtering,
  occurence_min_confidence,
  birdnet_params_list = list(),
  verbose = TRUE,
  internal_pool = NULL,
  coordinates_decimal_places = 1L,
  tflite_num_threads = 1
)
}
\arguments{
\item{deployment_id}{Integer. ID of the deployment to process.}

\item{deployments}{Data frame / sf object. Deployment metadata.}

\item{audio_files}{Data frame. Audio files metadata.}

\item{temporal_filtering}{Logical. Whether to group audio files by week to facilitate temporal species filtering.}

\item{occurence_min_confidence}{Numeric. Minimum eBird occurrence confidence. If 0, all available species in Birdnet will be used}

\item{birdnet_params_list}{List. Optional parameters to override BirdNET defaults.}

\item{verbose}{Logical. Print progress messages.}

\item{internal_pool}{DBI pool object. If NULL, a new pool will be created.}

\item{coordinates_decimal_places}{integer, number of decimal places for epsg4326 coordinates (only 1 or 2 decimal places are meaningful, 2 may lead to more fine grained species filters but 1 should be enough and reduce the number of settings in the database)}
}
\value{
Data frame of BirdNET inference results.
}
\description{
Main entry point for processing a deployment:
- filters audio files
- sets up BirdNET model
- generates settings
- applies inference
- aggregates results
}
