% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_manage_users.R, R/fct_user_editor.R
\name{update_user}
\alias{update_user}
\title{Update an Existing User}
\usage{
update_user(
  pool,
  username,
  password = NULL,
  pg_role = NULL,
  active = NULL,
  first_name = NULL,
  last_name = NULL,
  email = NULL,
  expire_date = NULL
)

update_user(
  pool,
  username,
  password = NULL,
  pg_role = NULL,
  active = NULL,
  first_name = NULL,
  last_name = NULL,
  email = NULL,
  expire_date = NULL
)
}
\arguments{
\item{pool}{A DBI connection pool.}

\item{username}{Character. Username whose record is updated.}

\item{password}{Optional character. New password.}

\item{pg_role}{Optional character. New PostgreSQL role to add.}

\item{active}{Optional logical.}

\item{first_name}{Optional character.}

\item{last_name}{Optional character.}

\item{email}{Optional character.}

\item{expire_date}{Optional Date.}
}
\value{
Invisible list with success status.

Invisible list with success status.
}
\description{
Allows updating of:
- password (rehashes automatically)
- active status
- first/last name
- email
- expire_date
- roles (adds new role; does not remove existing roles)

Allows updating of:
- password (rehashes automatically)
- active status
- first/last name
- email
- expire_date
- roles (adds new role; does not remove existing roles)
}
